backend/requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
pillow
sentence-transformers
nltk
faiss-cpu
requests
easyocr
transformers 
torch
rank_bm25